Eliminate signed/unsigned comparison

+++ testing/bulk_async/void_result.cpp
@@ -19,7 +19,7 @@ void test(ExecutionPolicy policy)
 
     f.wait();
 
-    size_t num_agents = agent_traits::domain(policy.param()).size();
+    int num_agents = agent_traits::domain(policy.param()).size();
 
     assert(counter == num_agents);
   }
@@ -41,7 +41,7 @@ void test(ExecutionPolicy policy)
 
     f.wait();
 
-    size_t num_agents = agent_traits::domain(policy.param()).size();
+    int num_agents = agent_traits::domain(policy.param()).size();
 
     assert(counter == num_agents * 13);
   }
@@ -63,7 +63,7 @@ void test(ExecutionPolicy policy)
 
     f.wait();
 
-    size_t num_agents = agent_traits::domain(policy.param()).size();
+    int num_agents = agent_traits::domain(policy.param()).size();
 
     assert(counter == num_agents * 13);
   }
